Re-factor auth logic
